Add files via upload
This is an Excel project that utilizes raw data from a variety of CSV files and then uses an ETL process in Power Query and Power Pivot to build a detailed pivot table of various customers and their net sales figures. The second spreadsheet displays the country and its associated net sales figures.